i7291: support large pages in TLB
Changed `block_size` to `long int` in order to allow larger pages in
TLB. Also added two minor fixes that came up while working on the
feature:
 - Argument name mismatches in `cache_lru_t` and `cache_fifo_t`
between the header and body.
 - Clearer error messages in `tlb_simulator_t`, so you can know
exactly what failed to build.

Fixes #7291

+++ clients/drcachesim/simulator/caching_device.cpp
@@ -41,9 +41,9 @@
 #include <utility>
 #include <vector>
 
-#include "memref.h"
 #include "caching_device_block.h"
 #include "caching_device_stats.h"
+#include "memref.h"
 #include "prefetcher.h"
 #include "snoop_filter.h"
 #include "trace_entry.h"
@@ -74,7 +74,7 @@ caching_device_t::~caching_device_t()
 }
 
 bool
-caching_device_t::init(int associativity, int block_size, int num_blocks,
+caching_device_t::init(int associativity, long int block_size, int num_blocks,
                        caching_device_t *parent, caching_device_stats_t *stats,
                        prefetcher_t *prefetcher,
                        cache_inclusion_policy_t inclusion_policy, bool coherent_cache,
